+++ docs/RELEASE_NOTES.md
@@ -1,3 +1,24 @@
+# Version 8.0.1 (12/10/2024)
+Added HTTP Api REST endpoints allowing to query: \
+-  Performance Monitoring Tool sensors configuration in use \
+-  Timestamp of the latest metadata local cache refresh \
+Added plugin for generating Promtheus config file authomatically based on the actual Performance Monitoring Tool sensors configuration \
+Added GPFSmmhealth sensor to the PrometheusExporter supported endpoints \
+Improved the  performance of the OpenTSDB HTTP Api REST  search/loookup endpoint \
+Added monitoring thread  observing if MetaData refresh is required \
+Published example dashboards showing: \
+- GPFS Cluster overview using OpenTSDB Datasource \
+- GPFS Cluster overview using Prometheus Datasource \
+- GPFSmmhealth metrics \
+- GPFS physical disks wait times \
+- GPFS fileset quota reporting \
+Changed the Dockerfile parent image to the registry.access.redhat.com/ubi9/ubi:9.5-1732804088 \
+
+Tested with Grafana version 11.0.0
+Tested with RedHat community-powered Grafana operator v.5
+
+
+
 # Version 8.0.0 (04/26/2024)
 The Grafana Bridge has been refactored to allow several APIs to be registered and run as standalone plugins. The OpenTSDB API now needs to be explicitly registered via port configuration in config.ini before it can be used with Grafana. 
 Added the new Prometheus Exporter plugin which collects metrics and exposes them in a format that can be scraped by the Prometheus timeseries database. This plugin also needs to be enabled via port configuration. 
@@ -6,6 +27,64 @@ Added a Client Basic Authentication over HTTP/S support.
 Added features to collect and report the internal performance statistics of grafana-bridge. 
 
 Tested with Grafana version 10.2.3
+Tested with RedHat community-powered Grafana operator v.5
+
+
+
+# Version 7.2.0 (12/06/2024)
+Changed the Dockerfile parent image to the registry.access.redhat.com/ubi9/ubi:9.5-1732804088 \
+Speed up OpenTSDB /search/lookup REST Api endpoint response time \
+Backported important fixes that improve overall bridge performance \
+
+Tested with Grafana version 11
+Tested with RedHat community-powered Grafana operator v.5
+
+
+
+# Version 7.1.9 (09/27/2024)
+Changed the Dockerfile parent image to the registry.access.redhat.com/ubi9/ubi:9.4-1214.1726694543 \
+
+Tested with Grafana version 11
+Tested with RedHat community-powered Grafana operator v.5
+
+
+
+# Version 7.1.8 (08/29/2024)
+Changed the Dockerfile parent image to the registry.access.redhat.com/ubi9/ubi:9.4-1181.1724035907 \
+
+Tested with Grafana version 11
+Tested with RedHat community-powered Grafana operator v.5
+
+
+
+# Version 7.1.7 (08/27/2024)
+Changed the Dockerfile parent image to the registry.access.redhat.com/ubi9/ubi:9.4-1181 \
+
+Tested with Grafana version 11
+Tested with RedHat community-powered Grafana operator v.5
+
+
+
+# Version 7.1.6 (06/28/2024)
+Changed the Dockerfile parent image to the registry.access.redhat.com/ubi9/ubi:9.4-1123 \
+
+Tested with Grafana version 11
+Tested with RedHat community-powered Grafana operator v.5
+
+
+
+# Version 7.1.5 (06/07/2024)
+Changed the Dockerfile parent image to the registry.access.redhat.com/ubi9/ubi:9.4-947.1717074712 \
+
+Tested with Grafana version 9.5
+Tested with RedHat community-powered Grafana operator v.5
+
+
+
+# Version 7.1.4 (05/07/2024)
+Changed the Dockerfile parent image to the registry.access.redhat.com/ubi9/ubi:9.4 \
+
+Tested with Grafana version 9.5
 Tested with RedHat community-powered Grafana operator v.5